chore(maintainability): retire turn bridge giant caps
